Fix issue in packtop and add finish top20.json generation
The issue was that I was creating a list of empty lists incorrectly, which made all the sub lists linked to the same reference.
This would have meant that as soon as toppack would encounter a case where more than one pack was owned by an individual, it would fail.
The same way of json-ifying the array for toppack is now being used for top20.
Had to add an extra function for helping to merge the unicodes for stickers into the data array.
Unfortunately the only way I can get the unicodes for now is by raising an error via print(msg) and parsing the error message for the unicode.